# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CollabLearn Unified API
  version: 1.0.0
  description: Shared API contract for CollabLearn web, Android, and iOS companion applications.
servers:
  - url: http://localhost:5001/api
    description: Local development backend
  - url: https://collablearn.vercel.app/api
    description: Cloud staging & production
paths:
...
  /booking:
    get:
      summary: List bookings for current user
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
    post:
      summary: Create new 1-on-1 mentor booking
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: collablearn_access_token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, mentor, admin]
        isSuperAdmin:
          type: boolean
        avatar:
          type: string
          nullable: true
        isPremium:
          type: boolean
...
    CreateBookingRequest:
      type: object
      required:
        - instructorId
        - skillId
        - date
        - startTime
        - endTime
      properties:
        instructorId:
          type: string
        skillId:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
    Booking:
      type: object
      properties:
        _id:
          type: string
        student:
          $ref: '#/components/schemas/User'
        instructor:
          $ref: '#/components/schemas/User'
        skill:
          type: object
        date:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
    BookingResponse:
      type: object
      properties:
        success:
          type: boolean
        booking:
          $ref: '#/components/schemas/Booking'
...